Update everything. Transfer of files from Cytometer computer to laptop. New scripts controlling the led lights. Lots of new images to detect Cytometer alerts. Improved clicker; added mouse movement before taking the screenshot to avoid capturing the black screen when capturing the screen. Complete_workflow.py adapted to new timing caused by change in the pumps (replaced the tape with filter paper to press the tube against the pump wall).

+++ clicker.py
@@ -1,21 +1,21 @@
 import CFlow,autopy,os,time,shutil,datetime
 from autopy import key
 
-
-
 class click:
     """Encapsulates some calls to the winapi for window management"""
-    def __init__(self,image_directory='C:/Users/rumarc/Desktop/Images'):
+    def __init__(self,sample_counter=0,image_directory='C:/Users/rumarc/Desktop/Images'):#Change directory where C6 images are stored
         """Constructor"""
         self._windowMgr=CFlow.WindowMgr()
         self._mouseMvr=CFlow.MouseMvr()
-        self.counter=0
+        self.time_counter=0
+        self.sample_counter=sample_counter
         self.image_directory=image_directory
+        self.export_counter=0
     def run(self,image_file="run_button_ready.png"):
         """Constructor"""
         self._windowMgr.retake_screenshot()
         try:
-            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#Name of the button image
+            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#path to the image
         except IOError:
             print('Image file not found. Quitting..')
             return
@@ -25,10 +25,12 @@ class click:
             position=(self._windowMgr._pos[0]+73,self._windowMgr._pos[1]+22)
             self._mouseMvr.move(position)
             self._mouseMvr.click()
-            self.counter=self.counter+1
-            self.delete_events(4,4)			#delete every for 4 seconds, 4 times. Delete the first 16 seconds of measurement.
+            self.time_counter=self.time_counter+1
+            self.sample_counter=self.sample_counter+1
+            self.delete_events()			#delete every 2 seconds, 5 times. Delete the first 10 seconds of measurement.
             if self.checking_end_of_measurements()==1:
-                self.add_sample_well_description() ##CHECK IF THIS INTERFERES WITH THE SAVING!
+                print('measurement %d done'%(self.time_counter))
+                self.add_sample_well_description()
                 self.save()
             else:
                 self.pause_cytometer()
@@ -36,46 +38,53 @@ class click:
                 self.save()
                 return						#When measurement takes more than 5minutes, pause operation and move on.
     def sample(self,image_file="sample_button2.png"):
+        self._windowMgr.maximize_window()
         """Constructor"""
         self._windowMgr.retake_screenshot()
         try:
-            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#Name of the button image
+            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#path to the image
         except IOError:
             print('Image file not found. Quitting..')
             return
         if self._windowMgr._pos == None:
             print('CFlow not open, no button positions available')
         else:
-            x_offset=(self.counter%12)*24
-            y_offset=(self.counter/12)*24
+            x_offset=(self.sample_counter%12)*24
+            y_offset=(self.sample_counter/12)*24
             position=(self._windowMgr._pos[0]+x_offset+36,self._windowMgr._pos[1]+y_offset+8)	#position changed to bring to center of button
             self._mouseMvr.move(position)
             self._mouseMvr.click()
     def save(self):   # It needs to have been saved once already!!
         """Constructor"""
-        self._windowMgr.maximize_window()
+#        self._windowMgr.maximize_window()		##REMOVE
+        key.tap(key.K_RETURN)
+        time.sleep(0.1)
         key.tap(key.K_ALT)
+        time.sleep(0.1)
+        for i in range(3):
+            key.tap(key.K_DOWN)
+            time.sleep(0.1)
         key.tap(key.K_RETURN)
-        key.tap(key.K_DOWN)
-        key.tap(key.K_DOWN)
-        key.tap(key.K_DOWN)
-        key.tap(key.K_RETURN)
-        #key.tap(key.K_RETURN); 
+        time.sleep(1)
     def export(self):
-        self._windowMgr.maximize_window()
-        key.tap(key.K_ALT)
+#        self._windowMgr.maximize_window()		##REMOVE
         key.tap(key.K_RETURN)
+        time.sleep(0.1)
+        key.tap(key.K_ALT)
+        time.sleep(0.1)
         for i in range(9):
             key.tap(key.K_DOWN)
+            time.sleep(0.1)
         key.tap(key.K_RETURN)
-        time.sleep(0.2)
+        time.sleep(0.3)
         key.tap(key.K_RETURN)
+        time.sleep(1)
 
     def backflush(self,image_file="backflush.png"):
         """Constructor"""
         self._windowMgr.retake_screenshot()
         try:
-            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#Name of the button image
+            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#path to the image
         except IOError:
             print('Image file not found. Quitting..')
             return
@@ -88,7 +97,7 @@ class click:
             if self.checking_end_of_measurements()==1:
                 return
             else:
-                return 'stuff took too long. Quitting..'
+                return 'backflushing took too long. Moving on..'
 
     def checking_end_of_measurements(self,image_file_ready="cytometer_ready.png"):##Add the checking whether CFlow is maximized. What if just don't minimize every time?
         """Constructor"""
@@ -98,42 +107,46 @@ class click:
         while not done:
             too_much_time=too_much_time+1
             time.sleep(5)
-            self._windowMgr.retake_screenshot()			##This causes CFlow to maximize its window every 5seconds, which is kind of annoying. Maybe just check if it is maximized?
+            self._windowMgr.retake_screenshot()
             try:
                 self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file_ready))	#Name of the button image
             except IOError:
                 print('Image file not found. Quitting..')
                 return
             if self._windowMgr._pos != None:
-                #print self._windowMgr._pos
-                print('measurements done')
                 done=1
                 return 1
             else:
-#                try:
-#                    self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file_not_ready))	#Name of the button image
-#                except IOError:
-#                    print('Image file not found. Quitting..')
-#                    return
-#                if self._windowMgr._pos == None:
-#                    
-#                    
-#                    
                 if too_much_time>60:					#When cytometer takes more than 5 minutes, pause operation and move on.
-                    print('Measurement takes too much time. Quitting..')
                     done=1
                     return 0
     def moveFiles(self,desktop_dir,destination_dir,CFlow_dir='FCS Exports'):
 ###move files from the folder that CFlow automatically makes to a specified folder.
         dir = os.path.join(desktop_dir,CFlow_dir)					#Find path to CFlow generated directory
-        automatic_CFlow_directory=[name for name in os.listdir(dir) if os.path.isdir(os.path.join(dir, name))]		#Find folder inside the directory. There should only be one.
-        automatic_CFlow_directory=os.path.join(dir,automatic_CFlow_directory[0])									#Get complete path to the folder where fcs files are
-        [shutil.copy2(os.path.join(automatic_CFlow_directory,name),destination_dir) for name in os.listdir(automatic_CFlow_directory)]	#move fcs files to new directory. If they are already present, they are replaced.
-        shutil.rmtree(automatic_CFlow_directory)																			#Remove the directory where the fcs files were. CHANGE TO SHUTIL.RMTREE()
-		####IT WORKS. TO TEST IT, ONE CAN USE THE FOLLOWING DIRECTORIES
-		##desktop_dir='C:\\Users\\rumarc\\Desktop'
-		##CFlow_dir='CFlow-FCS Exports'
-		##dir = os.path.join(desktop_dir,CFlow_dir)
+        if os.path.exists(dir) and os.path.exists(destination_dir):
+            automatic_CFlow_directory=[name for name in os.listdir(dir) if os.path.isdir(os.path.join(dir, name))]		#Find folder inside the directory. There should only be one.
+            if len(automatic_CFlow_directory)>0:
+                self.export_counter=0
+                automatic_CFlow_directory=os.path.join(dir,automatic_CFlow_directory[0])									#Get complete path to the folder where fcs files are
+                if os.path.exists(automatic_CFlow_directory):
+                    [shutil.copy2(os.path.join(automatic_CFlow_directory,name),destination_dir) for name in os.listdir(automatic_CFlow_directory)]	#move fcs files to new directory. If they are already present, they are replaced.
+                    shutil.rmtree(automatic_CFlow_directory)																			#Remove the directory where the fcs files were. CHANGE TO SHUTIL.RMTREE()
+                else:
+                    print('Specified CFlow directory (where it exports) could not be found')
+            elif self.export_counter!=1:
+                print('Error. No files found to export. Trying to export once more')
+                self._windowMgr.maximize_window()
+                self.export_counter=1
+                self.export()
+                time.sleep(2)
+                self.moveFiles(desktop_dir,destination_dir,CFlow_dir)
+            else:
+                print('Error. Still no files found to export')
+                self.export_counter=0
+			#desktop_dir='C:\\Users\\rumarc\\Desktop'
+			#dir = os.path.join(desktop_dir,CFlow_dir)
+        else:
+            print('specified paths for Desktop or Output could not be found')
     def set_measuring_times(self,day,hour,minute,frequency,num_samples):#frequency is in minutes
         self.measuring_times=[]
         time_today=datetime.datetime.now()
@@ -145,39 +158,43 @@ class click:
         for i in range(len(self.measuring_times)):
             print ('\n %s \n'%(self.measuring_times[i]))
     def set_waiting_time(self):
-        if len(self.measuring_times)>self.counter:
-            self.waiting_time=(self.measuring_times[self.counter]-datetime.datetime.now()).total_seconds()
+        if len(self.measuring_times)>self.time_counter:
+            self.waiting_time=(self.measuring_times[self.time_counter]-datetime.datetime.now()).total_seconds()
             if self.waiting_time<0:
-                print('timepoint %s already passed. Waiting for the next one'%(str(self.measuring_times[self.counter]))) #If timepoint has passed, wait for the next one (or should we just measure right away?)
-                self.counter=self.counter+1
+                print('timepoint %s already passed. Waiting for the next one'%(str(self.measuring_times[self.time_counter]))) #If timepoint has passed, wait for the next one (or should we just measure right away?)
+                self.time_counter=self.time_counter+1
+                self.sample_counter=self.sample_counter+1
                 self.set_waiting_time()
             else:
                 return 0						#This indicates that experiment is NOT done, and that there is a next measurement.
         else:
             print('Experiment is done!!')
             return 1						#This indicates that experiment IS done. Close everything.
-    def delete_events(self,image_file="delete_events.png",frequency,repetitions):##Frequency is how many seconds to erase every time. Repetitions is how many times to erase
+    def delete_events(self,frequency=2,repetitions=5,image_file="delete_events.png"):##Frequency is how many seconds to erase every time. Repetitions is how many times to erase
         """Constructor"""
         self._windowMgr.retake_screenshot()
         try:
-            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#Name of the button image
+            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#path to the image
         except IOError:
             print('Image file not found. Quitting..')
             return
         if self._windowMgr._pos == None:
             print('Could not delete events. Button not found on screen')
         else:
+            time.sleep(15)
             for i in range(repetitions):
                 time.sleep(frequency)
-                position=(self._windowMgr._pos[0]+32,self._windowMgr._pos[1]+6) #add position offset
+                position=(self._windowMgr._pos[0]+32,self._windowMgr._pos[1]+15) #add position offset
                 self._mouseMvr.move(position)
+                time.sleep(0.1)
                 self._mouseMvr.click()
+                time.sleep(0.4)
                 key.tap(key.K_RETURN)
     def pause_cytometer(self,image_file="pause.png"):
         """Constructor"""
         self._windowMgr.retake_screenshot()
         try:
-            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#Name of the button image
+            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#path to the image
         except IOError:
             print('Image file not found. Quitting..')
             return
@@ -191,7 +208,7 @@ class click:
         """Constructor"""
         self._windowMgr.retake_screenshot()
         try:
-            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#Name of the button image
+            self._windowMgr.find_button_coordinates(os.path.join(self.image_directory,image_file))	#path to the image
         except IOError:
             print('Image file not found. Quitting..')
             return
@@ -201,17 +218,7 @@ class click:
             position=(self._windowMgr._pos[0]+80,self._windowMgr._pos[1]+8) #add position offset
             self._mouseMvr.move(position)
             self._mouseMvr.click()
+            time.sleep(0.5)
             sample_time=datetime.datetime.now()
-            sample_time=' '.join([str(sample_time.hour),str(sample_time.minute),str(sample_time.second)])
+            sample_time=''.join([str(sample_time.hour),'h',str(sample_time.minute),'m'])
             key.type_string(sample_time)
-
-#    def set_led_times(self,day,hour,minute,frequency,num_samples):#frequency is in minutes
-#        self.measuring_times=[]
-#        time_today=datetime.datetime.now()
-#        starting_time=datetime.datetime(time_today.year,time_today.month,day,hour,minute,second=0) #User input is only day,hour,minute,second
-#        number_samples=num_samples
-#        for i in range(number_samples):
-#            self.measuring_times.append(starting_time+datetime.timedelta(minutes=frequency*i))
-#        print('The measuring times will be:')
-#        for i in range(len(self.measuring_times)):
-#            print ('\n %s \n'%(self.measuring_times[i]))
